[feat]: 경로 생성 기능 추가

+++ backend/src/main/java/io/ssafy/p/j9b304/backend/domain/walk/dto/response/WalkGetListResponseDto.java
@@ -22,7 +22,7 @@ public class WalkGetListResponseDto {
     Integer time;
 
     @Schema(description = "이동 거리(km)")
-    Float distance;
+    Double distance;
 
     @Setter
     @Schema(description = "산책로 지도 이미지")
